next gen audioplayer

+++ app/src/main/java/ch/zhaw/bait17/audio_signal_processing_toolbox/PlaybackListener.java
@@ -10,6 +10,6 @@ public interface PlaybackListener {
 
     public void onCompletion();
 
-    public void onAudioDataReceived(float[] data);
+    public void onAudioDataReceived(short[] data);
 
 }
